+++ engine-tests/src/tests/erc20.rs
@@ -3,15 +3,17 @@ use crate::prelude::{Address, U256};
 use crate::utils::{
     self,
     solidity::erc20::{self, ERC20Constructor, ERC20},
-    Signer,
+    str_to_account_id, Signer,
 };
 use aurora_engine::engine::EngineErrorKind;
 use aurora_engine::parameters::TransactionStatus;
 use aurora_engine_sdk as sdk;
 use aurora_engine_types::account_id::AccountId;
+use aurora_engine_types::borsh::BorshSerialize;
 use aurora_engine_types::parameters::connector::{
     Erc20Identifier, Erc20Metadata, SetErc20MetadataArgs,
 };
+use aurora_engine_types::parameters::engine::SetOwnerArgs;
 use bstr::ByteSlice;
 use libsecp256k1::SecretKey;
 use std::str::FromStr;
@@ -285,6 +287,70 @@ fn test_erc20_get_and_set_metadata() {
     let result = runner.one_shot().call(
         "get_erc20_metadata",
         &caller,
+        serde_json::to_vec::<Erc20Identifier>(
+            &AccountId::from_str(token_account_id).unwrap().into(),
+        )
+        .unwrap(),
+    );
+    assert!(result.is_ok());
+
+    let metadata: Erc20Metadata =
+        serde_json::from_slice(&result.unwrap().return_data.as_value().unwrap()).unwrap();
+    assert_eq!(metadata, new_metadata);
+}
+
+#[test]
+fn test_erc20_get_and_set_metadata_by_owner() {
+    let mut runner = utils::deploy_runner();
+    let token_account_id = "token";
+    let erc20_address = runner.deploy_erc20_token(token_account_id);
+    let caller = runner.aurora_account_id.clone();
+
+    // Change the owner of the aurora contract
+    let new_owner = "new_owner";
+    let set_owner_args = SetOwnerArgs {
+        new_owner: str_to_account_id(new_owner),
+    };
+
+    let result = runner.call("set_owner", &caller, set_owner_args.try_to_vec().unwrap());
+    assert!(result.is_ok());
+
+    let caller = new_owner;
+
+    // Getting ERC-20 metadata by Address.
+    let result = runner.one_shot().call(
+        "get_erc20_metadata",
+        caller,
+        serde_json::to_vec::<Erc20Identifier>(&erc20_address.into()).unwrap(),
+    );
+
+    assert!(result.is_ok());
+
+    let metadata: Erc20Metadata =
+        serde_json::from_slice(&result.unwrap().return_data.as_value().unwrap()).unwrap();
+    assert_eq!(metadata, Erc20Metadata::default());
+
+    let new_metadata = Erc20Metadata {
+        name: "USD Token".to_string(),
+        symbol: "USDT".to_string(),
+        decimals: 20,
+    };
+
+    let result = runner.call(
+        "set_erc20_metadata",
+        caller,
+        serde_json::to_vec(&SetErc20MetadataArgs {
+            erc20_identifier: erc20_address.into(),
+            metadata: new_metadata.clone(),
+        })
+        .unwrap(),
+    );
+    assert!(result.is_ok());
+
+    // Getting ERC-20 metadata by NEP-141 account id.
+    let result = runner.one_shot().call(
+        "get_erc20_metadata",
+        caller,
         serde_json::to_vec::<Erc20Identifier>(
             &AccountId::from_str(token_account_id).unwrap().into(),
         )