+++ CHANGES.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.3.1] 2023-10-26
+
+### Fixes
+
+- The smart contract owner whose account id is not the same as the contract account id can set ERC-20 metadata
+  by [@aleksuss] ([#858]). 
+
+[#858]: https://github.com/aurora-is-near/aurora-engine/pull/858
+
 ## [3.3.0] 2023-10-23
 
 ### Changes
@@ -538,7 +547,8 @@ struct SubmitResult {
 
 ## [1.0.0] - 2021-05-12
 
-[Unreleased]: https://github.com/aurora-is-near/aurora-engine/compare/3.3.0...develop
+[Unreleased]: https://github.com/aurora-is-near/aurora-engine/compare/3.3.1...develop
+[3.3.1]: https://github.com/aurora-is-near/aurora-engine/compare/3.3.0...3.3.1
 [3.3.0]: https://github.com/aurora-is-near/aurora-engine/compare/3.2.0...3.3.0
 [3.2.0]: https://github.com/aurora-is-near/aurora-engine/compare/3.1.0...3.2.0
 [3.1.0]: https://github.com/aurora-is-near/aurora-engine/compare/3.0.0...3.1.0